backend/requirements.txt
fastapi
uvicorn
httpx==0.25.2
PyJWT==2.8.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1